+++ lib/portlayer/storage/config.go
@@ -32,5 +32,5 @@ type Configuration struct {
 	// refer to the datstore + path), valued by the datastores and the path.
 	VolumeLocations map[string]url.URL `vic:"0.1" scope:"read-only"`
 
-	ScratchSize int `vic:"0.1" scope:"read-only" key:"scratch_size"`
+	ScratchSize int64 `vic:"0.1" scope:"read-only" key:"scratch_size"`
 }
